% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse.R
\name{performs_nse}
\alias{performs_nse}
\title{Detect Whether a Function Performs Non-Standard Evaluation}
\usage{
performs_nse(...)
}
\arguments{
\item{...}{\R functions.}
}
\description{
Detect Whether a Function Performs Non-Standard Evaluation
}
