% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sourcetools.R
\name{validate_syntax}
\alias{validate_syntax}
\title{Find Syntax Errors}
\usage{
validate_syntax(string)
}
\arguments{
\item{string}{A character vector (of length one).}
}
\description{
Find syntax errors in a string of \R code.
}

