% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parsr.R
\name{tokenize_string}
\alias{tokenize_string}
\title{Tokenize an R String}
\usage{
tokenize_string(string)
}
\arguments{
\item{string}{A character vector (of length one).}
}
\description{
Read and tokenize an R string.
}

